# Optimize Seq and Map Readers with Iterator API