I've taken a look into issue #315 and noticed that random access with `PySequence_GetItem` was used to read individual items. Lists are generally slow with random access and the iterator protocol gives better performance when reading all items. So I've switched over to the iterator protocol for `Reader[Seq[_]]` and `Reader[Map[_, _]]`. A simple benchmark shows a speedup of 20-30%.

As more data structures support the iterator protocol than the random access protocol, it also allows the `Reader[Seq[_]]` to read more data structures, like sets. So this PR resolves issue #315.

<details>
<summary>Benchmark code</summary>

```
package me.shadaj.scalapy.py

import me.shadaj.scalapy.py
import org.scalatest.funsuite.AnyFunSuite

class ReadBenchmarkTest extends AnyFunSuite{
  test("benchmark seq"){
    val t0 = System.currentTimeMillis()
    val d = 0 until 10_000
    val ints = py.eval(d.mkString("[", ",", "]"))
    for(_ <- 0 until 1_000){
      ints.as[Seq[Int]]
    }
    val t1 = System.currentTimeMillis()
    info(s"${t1 - t0} ms")

    // new: 3418 ms
    // old: 4201 ms
  }

  test("benchmark map"){
    val t0 = System.currentTimeMillis()
    val d = (0 until 10_000).map(i => s"$i: $i")
    val map = py.eval(d.mkString("{", ",", "}"))
    for(_ <- 0 until 1_000){
      map.as[Map[Int, Int]]
    }
    val t1 = System.currentTimeMillis()
    info(s"${t1 - t0} ms")
    // new: 27923 ms
    // old: 38645 ms
  }
}
```
</details>