@shadaj 2024-08-28
> I've updated both statements as requested.
> 
> I must say, I don't fully understand why we need to decrease the references as we don't increase them, neither directly nor by calling `PyValue.fromBorrowed`. But it works this way, so I'm fine with it.

So the reason this is necessary is because the CPython APIs for getting an element from an iterator increments the reference count of that element before it's handed to us, rather than returning a borrowed value.